+++ src/pages/MyKiva/MyKivaNextStepsContent.vue
@@ -724,7 +724,7 @@ watch(() => props.goalRefreshKey, async (newVal, oldVal) => {
 });
 
 onMounted(async () => {
-	await checkCompletedGoal({ category: 'portfolio', persistHideGoalCard: true, cookieStore });
+	await checkCompletedGoal({ category: 'portfolio', persistHideGoalCard: true });
 	if (shouldShowPostLendingNextStepsCards) {
 		removePostLendingCardCookie(cookieStore);
 	}